+++ src/main/java/prismatic/shards/stellarity/registry/StellaritySoundEvents.java
@@ -28,6 +28,9 @@ public interface StellaritySoundEvents {
 
 	SoundEvent COPPER_ELEKTRA_SHIELD_DASH = register("item.copper_elektra_shield.dash");
 
+	SoundEvent VOIDED_ACTIVATE = register("effect.voided.activate");
+	SoundEvent VOIDED_DEACTIVATE = register("effect.voided.deactivate");
+
 	Holder<SoundEvent> FIRES_OF_HOKKAI = registerHolder("music_disc.fires_of_hokkai");
 	Holder<SoundEvent> DEVIANTS_LIGHT_MUSIC_BOX = registerHolder("music_disc.deviants_light_music_box");
 	Holder<SoundEvent> PRECIPICE_STEREO = registerHolder("music_disc.precipice_stereo");