+++ src/main/java/prismatic/shards/stellarity/registry/item/CopperElektraShield.java
@@ -6,18 +6,19 @@ import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.tags.DamageTypeTags;
 import net.minecraft.tags.ItemTags;
-import net.minecraft.world.entity.Entity;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.monster.Creeper;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.projectile.ProjectileUtil;
 import net.minecraft.world.item.DyeColor;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Rarity;
 import net.minecraft.world.item.ShieldItem;
 import net.minecraft.world.item.component.BlocksAttacks;
 import net.minecraft.world.level.ClipContext;
+import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BannerPatternLayers;
 import net.minecraft.world.level.block.entity.BannerPatternLayers.Layer;
 import net.minecraft.world.level.block.entity.BannerPatterns;
@@ -25,11 +26,9 @@ import net.minecraft.world.level.portal.TeleportTransition;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
 import org.jspecify.annotations.NonNull;
-import org.jspecify.annotations.Nullable;
 import prismatic.shards.stellarity.key.StellarityDamageTypes;
 import prismatic.shards.stellarity.networking.ClientboundElectricDashPayload;
 import prismatic.shards.stellarity.registry.StellarityDataComponents;
-import prismatic.shards.stellarity.registry.StellarityItems;
 import prismatic.shards.stellarity.util.tuple.Tuple2;
 
 import java.util.ArrayList;
@@ -77,46 +76,44 @@ public class CopperElektraShield extends ShieldItem {
 		.component(DataComponents.BREAK_SOUND, SoundEvents.SHIELD_BREAK);
 
 	@Override
-	public void inventoryTick(@NonNull ItemStack itemStack, @NonNull ServerLevel level, @NonNull Entity owner, @Nullable EquipmentSlot slot) {
-		super.inventoryTick(itemStack, level, owner, slot);
+	public @NonNull InteractionResult use(@NonNull Level level, @NonNull Player player, @NonNull InteractionHand hand) {
+		if (!(level instanceof ServerLevel serverLevel && player.isShiftKeyDown())) return super.use(level, player, hand);
 
-		if (!(owner instanceof Player player && player.isShiftKeyDown() && player.isUsingItem() && player.getUseItem().is(StellarityItems.COPPER_ELEKTRA_SHIELD)))
-			return;
-
-
+		var itemStack = player.getItemInHand(hand);
 		var time = level.getGameTime();
 		var rechargesAt = itemStack.get(StellarityDataComponents.RECHARGES_AT);
 		if (rechargesAt == null) rechargesAt = time;
-		else if (rechargesAt - 2 * DASH_CHARGE_TIME > time) return;
-		rechargesAt = Math.max(rechargesAt, time - 2 * DASH_CHARGE_TIME);
+		else if (rechargesAt - 2 * DASH_CHARGE_TIME > time) return InteractionResult.FAIL;
+		rechargesAt = Math.max(rechargesAt, time);
 
 		var ownerPosition = player.getEyePosition();
 		var proposedPosition = ownerPosition.add(player.getLookAngle().normalize().scale(7));
-		var raycast = level.clip(new ClipContext(ownerPosition, proposedPosition, ClipContext.Block.COLLIDER, ClipContext.Fluid.NONE, player));
+		var raycast = serverLevel.clip(new ClipContext(ownerPosition, proposedPosition, ClipContext.Block.COLLIDER, ClipContext.Fluid.NONE, player));
 		var endLocation = raycast.getLocation();
 		var entitiesHit = ProjectileUtil.getManyEntityHitResult(level, player, ownerPosition, endLocation, new AABB(ownerPosition, endLocation).inflate(2), (entity) -> entity instanceof LivingEntity target && player.canAttack(target), false);
 
-		var electric = level.damageSources().source(StellarityDamageTypes.ELECTRIC, owner, owner);
+		var electric = serverLevel.damageSources().source(StellarityDamageTypes.ELECTRIC, player, player);
 		List<Vec3> creeperLocations = new ArrayList<>();
 		for (var hitEntity : entitiesHit) {
 			var hit = hitEntity.getEntity();
-			hit.hurtServer(level, electric, 4);
+			hit.hurtServer(serverLevel, electric, 4);
 			if (hit instanceof Creeper creeper && creeper.getRandom().nextFloat() < 0.25F) {
 				creeper.getEntityData().set(Creeper.DATA_IS_POWERED, true);
 				creeperLocations.add(creeper.position());
 			}
 		}
 
+//		player.getCooldowns().addCooldown(itemStack, 5);
 
-		var simulationDistance = (level.getServer().getPlayerList().getSimulationDistance() + 1) * 16;
-		for (var networkPlayer : level.getPlayers(networkPlayer -> networkPlayer.position().distanceTo(endLocation) < simulationDistance)) {
+
+		var simulationDistance = (serverLevel.getServer().getPlayerList().getSimulationDistance() + 1) * 16;
+		for (var networkPlayer : serverLevel.getPlayers(networkPlayer -> networkPlayer.position().distanceTo(endLocation) < simulationDistance)) {
 			ServerPlayNetworking.send(networkPlayer, new ClientboundElectricDashPayload(ownerPosition, endLocation, creeperLocations));
 		}
 
-		owner.teleport(new TeleportTransition(level, endLocation, player.getDeltaMovement(), player.getYRot(), player.getXRot(), TeleportTransition.DO_NOTHING));
+		player.teleport(new TeleportTransition(serverLevel, endLocation, player.getDeltaMovement(), player.getYRot(), player.getXRot(), TeleportTransition.DO_NOTHING));
 
 		itemStack.set(StellarityDataComponents.RECHARGES_AT, rechargesAt + DASH_CHARGE_TIME);
-
-
+		return InteractionResult.SUCCESS_SERVER;
 	}
 }